splitted handler into a separate file
